电池查询带时间

+++ src/main/java/com/xiliulou/electricity/entity/BatteryAlert.java
@@ -36,7 +36,7 @@ public class BatteryAlert {
 	/**
 	 * 	告警标识
 	 */
-	private Integer alarmFlag;
+	private Long alarmFlag;
 	/**
 	 * 	type
 	 */
